scheduled_messages: Move update_send_later_options.
